flaskapp/requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
pandas==0.25.3
plotly==4.4.1
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
retrying==1.3.3
six==1.13.0
Werkzeug==0.16.0
